client/server update für die stabilität

+++ web/download.php
@@ -1,16 +1,19 @@
 <?php
 	$name=$_GET['name'];
 	$path=$_GET['path'];
+	$del=$_GET['del'];
 	if($path=='tmp'){
-	$pfad = "/var/www/tmp/".$name;
+		$pfad = "/var/www/tmp/".$name;
 	}
 	else{
-	$pfad = "/var/www/save/".$name;
+		$pfad = "/var/www/save/".$name;
 	}
 	header("Content-Type: application/force-download");
 	header("Content-Disposition: attachment; filename=\"". $name ."\"");
 	header("Content-Length: ". filesize($pfad));
 	header("Content-Transfer-Encoding: binary");
 	readfile($pfad);
-	unlink($pfad);
+	if($del=='del'){
+		unlink($pfad);
+	}
 ?>
